requirements.txt
asgiref==3.8.1
Django==5.0.7
django-bootstrap5==24.2
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
